// packages/cosapp-lab/cosapp_lab-0.15.0.tar.gz/cosapp_lab-0.15.0/src/sys_explorer_react/redux/types.ts
export interface StateInterface {
  /** Global configuration */
  systemConfig: { [key: string]: any };
  /** State of activated panel */
  mainState: PanelStatusInterface;
  saveSignal: number;
  /** State of system architecture panel*/
  systemArch: {
    lockStatus: boolean; // lock status of diagram
    /** State of system tree structure */
    systemTree: {
      nodeData: any; // Data of each node
      nodePath: string[]; // Path from root to each node
      selectedNode: string; // Current selected node for editing
    };
    /** State which holds the data of each system */
    systemData: SystemDataInterface;
    /** State which holds the data of for diagram */
    systemGraph: SystemGraphInterface;
    /** State of system pbs view */
    systemPBS: { [key: string]: { visible: boolean; position: Array<number> } };
    /**Signal for node filtering */
    systemPBSUpdated: number;
  };
  /**State of dashboard panel */
  dashboardState: {
    computing: boolean; // True if a driver is running, False otherwise.
    timeStepList: string[]; // List of time step available.
    data3D: { [key: string]: any[] }; // Not used yet.
    variableData: { [key: string]: any }; // System variable data received from backend.
    portMetaData: { [key: string]: any }; // Port meta data received from backend.
    selectedVariable: { [key: string]: number }; // Controller variable and its value.
    computedResult: { [key: string]: any }; // Value of all variable, used for plot
    recorderData: { [key: string]: any }; // Cosapp recorder
    driverData: { [key: string]: any }; // Cosapp driver data
  };
}
export interface SystemDataInterface {
  mainData: {
    // key is name of system
    [key: string]: {
      user: UserListType[];
      class: string;
      portList: string[];
    };
  };
  portData: {
    //key in form of "systemName/portName"
    [key: string]: {
      variable: string[];
      connection: string[]; // Destination port in form of "systemName/portName"
    };
  };
}

export interface PanelStatusInterface {
  [key: string]: [boolean, string];
}

export interface SystemGraphInterface {
  systemGraphData: {
    [key: string]: {
      inPort: Array<string>;
      outPort: Array<string>;
      connections: Array<Array<any>>;
      position?: Array<number>;
    };
  };
  systemList: Array<string>;
  graphJsonData: { [key: string]: any };
  updateData: { [key: string]: any };
}
// ...
export interface UserListType {
  uid: number;
  name: string;
}